Rename Shadowsocks.Proxy -> Shadowsocks.ForwardProxy
